modules/location_event/requirements.txt
grpcio-tools
grpcio
kafka-python